Harden Android Keycloak configuration (#204)

+++ android/app/src/main/java/com/mavi/vpn/ui/components/MaviCore.kt
@@ -164,7 +164,7 @@ fun MaviCore(
                 val dotX = cx + (cos(a) * orbitR).toFloat()
                 val dotY = cy + (sin(a) * orbitR).toFloat()
                 val dotR = if (i == 0) 4f.dp.toPx() else 2f.dp.toPx()
-                val opacity = if (i == 0) 1f else 0.4f + 0.3f * sin(timeSeconds * 2f + i).toFloat()
+                val opacity = if (i == 0) 1f else 0.4f + 0.3f * sin(timeSeconds * 2f + i)
                 
                 drawCircle(
                     color = accent.copy(alpha = opacity),
